Moving ledger into appState part 2
