+++ app/renderer/components/preferences/payment/ledgerTable.js
@@ -24,6 +24,7 @@ const pinIcon = require('../../../../extensions/brave/img/ledger/icon_pin.svg')
 const settings = require('../../../../../js/constants/settings')
 const getSetting = require('../../../../../js/settings').getSetting
 const aboutActions = require('../../../../../js/about/aboutActions')
+const urlUtil = require('../../../../../js/lib/urlutil')
 const {SettingCheckbox, SiteSettingCheckbox} = require('../../common/settings')
 
 class LedgerTable extends ImmutableComponent {
@@ -51,7 +52,7 @@ class LedgerTable extends ImmutableComponent {
   }
 
   getHostPattern (synopsis) {
-    return `https?://${synopsis.get('site')}`
+    return urlUtil.getHostPattern(synopsis.get('site'))
   }
 
   getVerifiedIcon (synopsis) {
