+++ js/constants/appConstants.js
@@ -33,7 +33,6 @@ const appConstants = {
   APP_ON_CLEAR_BROWSING_DATA: _,
   APP_IMPORT_BROWSER_DATA: _,
   APP_UPDATE_LEDGER_INFO: _,
-  APP_UPDATE_LOCATION_INFO: _,
   APP_SHOW_NOTIFICATION: _, /** @param {Object} detail */
   APP_HIDE_NOTIFICATION: _, /** @param {string} message */
   APP_BACKUP_KEYS: _,
@@ -153,7 +152,24 @@ const appConstants = {
   APP_INSPECT_ELEMENT: _,
   APP_ON_BOOKMARK_WIDTH_CHANGED: _,
   APP_ON_BOOKMARK_FOLDER_WIDTH_CHANGED: _,
-  APP_WINDOW_RESIZED: _
+  APP_WINDOW_RESIZED: _,
+  APP_ON_FAVICON_RECEIVED: _,
+  APP_ON_PUBLISHER_OPTION_UPDATE: _,
+  APP_ON_LEDGER_OPTION_UPDATE: _,
+  APP_ON_LEDGER_WALLET_CREATE: _,
+  APP_ON_BOOT_STATE_FILE: _,
+  APP_LEDGER_PAYMENTS_PRESENT: _,
+  APP_ON_WALLET_RECOVERY: _,
+  APP_ON_BRAVERY_PROPERTIES: _,
+  APP_ON_LEDGER_BALANCE_RECEIVED: _,
+  APP_ON_LEDGER_LOCATION_UPDATE: _,
+  APP_ON_WALLET_PROPERTIES: _,
+  APP_ON_ADD_FUNDS_CLOSED: _,
+  APP_ON_FIRST_LEDGER_SYNC: _,
+  APP_ON_LEDGER_CALLBACK: _,
+  APP_ON_TIME_UNTIL_RECONCILE: _,
+  APP_ON_LEDGER_RUN: _,
+  APP_ON_NETWORK_CONNECTED: _
 }
 
 module.exports = mapValuesByKeys(appConstants)
