+++ js/constants/messages.js
@@ -132,12 +132,9 @@ const messages = {
   // Debugging
   DEBUG_REACT_PROFILE: _,
   // Ledger
-  LEDGER_PAYMENTS_PRESENT: _,
   LEDGER_PUBLISHER: _,
   LEDGER_PUBLISHER_RESPONSE: _,
   LEDGER_UPDATED: _,
-  LEDGER_CREATE_WALLET: _,
-  ADD_FUNDS_CLOSED: _,
   RENDER_URL_TO_PDF: _,
   // Sync
   SYNC_UPDATED: _,
