+++ test/lib/brave.js
@@ -1052,7 +1052,7 @@ var exports = {
       return this.waitUntil(function () {
         return this.getAppState().then((val) => {
           val = Immutable.fromJS(val)
-          let synopsis = val.getIn(['value', 'publisherInfo', 'synopsis'])
+          let synopsis = val.getIn(['value', 'ledger', 'synopsis'])
           if (synopsis !== undefined) {
             return cb(synopsis)
           }
